requirements.txt
flask==2.3.2
flask-socketio==5.3.0
python-socketio==5.8.0
eventlet==0.33.3
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.0
Pillow==9.5.0
perplexityai==0.11.0
gunicorn==21.2.0
gevent==23.9.1
gevent-websocket==0.10.1

